requirements.txt
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
rasterio>=1.3.0
requests>=2.31.0
h3>=3.7.0
boto3>=1.28.0

# Note: Agisoft Metashape Python API must be installed separately
# It is not available via pip. Install from: https://www.agisoft.com/downloads/installer/
